requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-multipart>=0.0.9
sqlalchemy[asyncio]>=2.0
aiosqlite>=0.19
pydantic>=2.6
jinja2>=3.1
python-dotenv>=1.0
openai>=1.0
instructor>=1.0
openpyxl>=3.1
markitdown>=0.0.1
